Separated tests for headers and httpClient
