+++ lib/httpClient.js
@@ -1,28 +1,19 @@
-'use strict';
-
-/**
- * Load Test a URL, website or websocket.
- * (C) 2013 Alex Fernández.
- */
-
-
-const testing = require('testing');
-const urlLib = require('url');
-const http = require('http');
-const https = require('https');
-const qs = require('querystring');
-const websocket = require('websocket');
-const {HighResolutionTimer} = require('./hrtimer.js');
-const headers = require('./headers.js');
+import * as urlLib from 'url'
+import * as http from 'http'
+import * as https from 'https'
+import * as querystring from 'querystring'
+import * as websocket from 'websocket'
+import {HighResolutionTimer} from './hrtimer.js'
+import {addUserAgent} from './headers.js'
 
 
 /**
  * Create a new HTTP client.
  * Seem parameters below.
  */
-exports.create = function(operation, params) {
+export function create(operation, params) {
 	return new HttpClient(operation, params);
-};
+}
 
 /**
  * A client for an HTTP connection.
@@ -72,7 +63,7 @@ class HttpClient {
 				this.generateMessage = () => this.params.body;
 			} else if (typeof this.params.body == 'object') {
 				if (this.params.contentType === 'application/x-www-form-urlencoded') {
-					this.params.body = qs.stringify(this.params.body);
+					this.params.body = querystring.stringify(this.params.body);
 				}
 				this.generateMessage = () => this.params.body;
 			} else if (typeof this.params.body == 'function') {
@@ -91,7 +82,7 @@ class HttpClient {
 				console.error('Invalid cookies %j, please use an array or a string', this.params.cookies);
 			}
 		}
-		headers.addUserAgent(this.options.headers);
+		addUserAgent(this.options.headers);
 		if (this.params.secureProtocol) {
 			this.options.secureProtocol = this.params.secureProtocol;
 		}
@@ -259,30 +250,5 @@ class HttpClient {
 			});
 		};
 	}
-}
-
-function testHttpClient(callback) {
-	const options = {
-		url: 'http://localhost:7357/',
-		maxSeconds: 0.1,
-		concurrency: 1,
-	};
-	exports.create({}, options);
-	testing.success(callback);
-}
-
-
-/**
- * Run all tests.
- */
-exports.test = function (callback) {
-	testing.run([
-		testHttpClient,
-	], callback);
-};
-
-// run tests if invoked directly
-if (__filename == process.argv[1]) {
-	exports.test(testing.show);
 }
