+++ test/run.js
@@ -5,13 +5,15 @@
 
 import testing from 'testing'
 import {test as testHrtimer} from './hrtimer.js'
+import {test as testHeaders} from './headers.js'
+import {test as testHttpClient} from './httpClient.js'
 
 
 /**
  * Run all module tests.
  */
 function test() {
-	const tests = [testHrtimer];
+	const tests = [testHrtimer, testHeaders, testHttpClient];
 	testing.run(tests, 4200);
 }
